Move timer definitions from quda_internal.h into timer.h and add dummy implementation of TimeProfile that is more robustly jit safe.  Only include assert.h if not using nvrtc
